+++ src/ballistica/base/audio/audio_server.cc
@@ -47,8 +47,12 @@ extern std::string g_rift_audio_device_name;
 LPALCDEVICEPAUSESOFT alcDevicePauseSOFT{};
 LPALCDEVICERESUMESOFT alcDeviceResumeSOFT{};
 LPALCRESETDEVICESOFT alcResetDeviceSOFT{};
+LPALCREOPENDEVICESOFT alcReopenDeviceSOFT{};
 LPALEVENTCALLBACKSOFT alEventCallbackSOFT{};
 LPALEVENTCONTROLSOFT alEventControlSOFT{};
+LPALCEVENTCALLBACKSOFT alcEventCallbackSOFT{};
+LPALCEVENTCONTROLSOFT alcEventControlSOFT{};
+LPALCEVENTISSUPPORTEDSOFT alcEventIsSupportedSOFT{};
 // LPALSOFTSETLOGCALLBACK alsoft_set_log_callback{};
 
 #endif
@@ -178,6 +182,15 @@ static void ALEventCallback_(ALenum eventType, ALuint object, ALuint param,
                              + std::to_string(static_cast<int>(eventType)));
   }
 }
+static void ALCEventCallback_(ALCenum eventType, ALCenum deviceType,
+                              ALCdevice* eventDevice, ALCsizei msgLen,
+                              const ALCchar* msg, void* user) noexcept {
+  if (eventType == ALC_EVENT_TYPE_DEFAULT_DEVICE_CHANGED_SOFT
+      && deviceType == ALC_PLAYBACK_DEVICE_SOFT) {
+    g_base->audio_server->event_loop()->PushCall(
+        [] { g_base->audio_server->OnDefaultDeviceChanged(); });
+  }
+}
 
 // FIXME: Should convert this to a generalized OpenALSoft log handler since
 // we might want to wire it up on other platforms too.
@@ -187,6 +200,7 @@ static void ALCustomAndroidLogCallback_(int severity, const char* msg) {
   // everything else (up to some size limit). We can then explicitly ship
   // the full log if a serious problem occurs.
   if (severity >= ANDROID_LOG_WARN) {
+    // Log(LogName::kAudio, LogLevel::k
     __android_log_print(severity, "BallisticaKit", "openal-log: %s", msg);
   }
   g_base->audio_server->OpenALSoftLogCallback(msg);
@@ -416,35 +430,60 @@ void AudioServer::Start_() {
     CHECK_AL_ERROR;
 
 #if BA_OPENAL_IS_SOFT
-    // Currently assuming the pause/resume and reset extensions are present.
-    // if (alcIsExtensionPresent(device, "ALC_SOFT_pause_device")) {
+    // Currently assuming these extensions are present.
     alcDevicePauseSOFT = reinterpret_cast<LPALCDEVICEPAUSESOFT>(
         alcGetProcAddress(device, "alcDevicePauseSOFT"));
-    BA_PRECONDITION_FATAL(alcDevicePauseSOFT != nullptr);
     alcDeviceResumeSOFT = reinterpret_cast<LPALCDEVICERESUMESOFT>(
         alcGetProcAddress(device, "alcDeviceResumeSOFT"));
-    BA_PRECONDITION_FATAL(alcDeviceResumeSOFT != nullptr);
+
+    // We expect neither or both of these to be present.
+    assert((alcDevicePauseSOFT != nullptr) == (alcDeviceResumeSOFT != nullptr));
+
     alcResetDeviceSOFT = reinterpret_cast<LPALCRESETDEVICESOFT>(
         alcGetProcAddress(device, "alcResetDeviceSOFT"));
     BA_PRECONDITION_FATAL(alcResetDeviceSOFT != nullptr);
+    alcReopenDeviceSOFT = reinterpret_cast<LPALCREOPENDEVICESOFT>(
+        alcGetProcAddress(device, "alcReopenDeviceSOFT"));
+    BA_PRECONDITION_FATAL(alcReopenDeviceSOFT != nullptr);
     alEventCallbackSOFT = reinterpret_cast<LPALEVENTCALLBACKSOFT>(
         alcGetProcAddress(device, "alEventCallbackSOFT"));
     BA_PRECONDITION_FATAL(alEventCallbackSOFT != nullptr);
     alEventControlSOFT = reinterpret_cast<LPALEVENTCONTROLSOFT>(
         alcGetProcAddress(device, "alEventControlSOFT"));
     BA_PRECONDITION_FATAL(alEventControlSOFT != nullptr);
-    // alsoft_set_log_callback = reinterpret_cast<LPALSOFTSETLOGCALLBACK>(
-    //     alcGetProcAddress(device, "alsoft_set_log_callback"));
-    // BA_PRECONDITION_FATAL(alsoft_set_log_callback != nullptr);
+    alcEventCallbackSOFT = reinterpret_cast<LPALCEVENTCALLBACKSOFT>(
+        alcGetProcAddress(device, "alcEventCallbackSOFT"));
+    BA_PRECONDITION_FATAL(alcEventCallbackSOFT != nullptr);
+    alcEventControlSOFT = reinterpret_cast<LPALCEVENTCONTROLSOFT>(
+        alcGetProcAddress(device, "alcEventControlSOFT"));
+    BA_PRECONDITION_FATAL(alcEventControlSOFT != nullptr);
 
-    // Ask to be notified when a device is disconnected.
+    alcReopenDeviceSOFT = reinterpret_cast<LPALCREOPENDEVICESOFT>(
+        alcGetProcAddress(device, "alcReopenDeviceSOFT"));
+    BA_PRECONDITION_FATAL(alcReopenDeviceSOFT != nullptr);
+
+    alcEventIsSupportedSOFT = reinterpret_cast<LPALCEVENTISSUPPORTEDSOFT>(
+        alcGetProcAddress(device, "alcEventIsSupportedSOFT"));
+
+    // Wire up our event callback.
     alEventCallbackSOFT(ALEventCallback_, nullptr);
     CHECK_AL_ERROR;
+    alcEventCallbackSOFT(ALCEventCallback_, nullptr);
+    CHECK_AL_ERROR;
+
+    // Ask to be notified when a device is disconnected.
     ALenum types[] = {AL_EVENT_TYPE_DISCONNECTED_SOFT};
     alEventControlSOFT(1, types, AL_TRUE);
-    // } else {
-    //   FatalError("ALC_SOFT pause/resume functionality not found.");
-    // }
+    CHECK_AL_ERROR;
+
+    if (alcEventIsSupportedSOFT(ALC_EVENT_TYPE_DEFAULT_DEVICE_CHANGED_SOFT,
+                                ALC_PLAYBACK_DEVICE_SOFT)
+        == ALC_EVENT_SUPPORTED_SOFT) {
+      // Ask to be notified when default output device changes.
+      ALenum types[] = {ALC_EVENT_TYPE_DEFAULT_DEVICE_CHANGED_SOFT};
+      alcEventControlSOFT(1, types, AL_TRUE);
+      CHECK_AL_ERROR;
+    }
 #endif
   }
 
@@ -561,27 +600,30 @@ void AudioServer::SetSuspended_(bool suspend) {
 
       // Pause OpenALSoft.
 #if BA_OPENAL_IS_SOFT
-      BA_PRECONDITION_FATAL(alcDevicePauseSOFT != nullptr);
-      BA_PRECONDITION_FATAL(impl_ != nullptr && impl_->alc_context != nullptr);
-      auto* device = alcGetContextsDevice(impl_->alc_context);
-      BA_PRECONDITION_FATAL(device != nullptr);
+      if (alcDevicePauseSOFT != nullptr) {
+        BA_PRECONDITION_FATAL(impl_ != nullptr
+                              && impl_->alc_context != nullptr);
+        auto* device = alcGetContextsDevice(impl_->alc_context);
+        BA_PRECONDITION_FATAL(device != nullptr);
 
-      try {
-        g_core->platform->LowLevelDebugLog(
-            "Calling alcDevicePauseSOFT at "
-            + std::to_string(g_core->AppTimeSeconds()));
-        alcDevicePauseSOFT(device);
-      } catch (const std::exception& e) {
-        g_core->logging->Log(
-            LogName::kBaAudio, LogLevel::kError,
-            "Error in alcDevicePauseSOFT at time "
-                + std::to_string(g_core->AppTimeSeconds()) + "( playing since "
-                + std::to_string(last_started_playing_time_)
-                + "): " + g_core->platform->DemangleCXXSymbol(typeid(e).name())
-                + " " + e.what());
-      } catch (...) {
-        g_core->logging->Log(LogName::kBaAudio, LogLevel::kError,
-                             "Unknown error in alcDevicePauseSOFT");
+        try {
+          g_core->platform->LowLevelDebugLog(
+              "Calling alcDevicePauseSOFT at "
+              + std::to_string(g_core->AppTimeSeconds()));
+          alcDevicePauseSOFT(device);
+        } catch (const std::exception& e) {
+          g_core->logging->Log(
+              LogName::kBaAudio, LogLevel::kError,
+              "Error in alcDevicePauseSOFT at time "
+                  + std::to_string(g_core->AppTimeSeconds())
+                  + "( playing since "
+                  + std::to_string(last_started_playing_time_) + "): "
+                  + g_core->platform->DemangleCXXSymbol(typeid(e).name()) + " "
+                  + e.what());
+        } catch (...) {
+          g_core->logging->Log(LogName::kBaAudio, LogLevel::kError,
+                               "Unknown error in alcDevicePauseSOFT");
+        }
       }
 #endif
 
@@ -605,25 +647,27 @@ void AudioServer::SetSuspended_(bool suspend) {
 
 // With OpenALSoft lets tell openal-soft to resume processing.
 #if BA_OPENAL_IS_SOFT
-      BA_PRECONDITION_FATAL(alcDeviceResumeSOFT != nullptr);
-      BA_PRECONDITION_FATAL(impl_ != nullptr && impl_->alc_context != nullptr);
-      auto* device = alcGetContextsDevice(impl_->alc_context);
-      BA_PRECONDITION_FATAL(device != nullptr);
-      try {
-        g_core->platform->LowLevelDebugLog(
-            "Calling alcDeviceResumeSOFT at "
-            + std::to_string(g_core->AppTimeSeconds()));
-        alcDeviceResumeSOFT(device);
-      } catch (const std::exception& e) {
-        g_core->logging->Log(
-            LogName::kBaAudio, LogLevel::kError,
-            "Error in alcDeviceResumeSOFT at time "
-                + std::to_string(g_core->AppTimeSeconds()) + ": "
-                + g_core->platform->DemangleCXXSymbol(typeid(e).name()) + " "
-                + e.what());
-      } catch (...) {
-        g_core->logging->Log(LogName::kBaAudio, LogLevel::kError,
-                             "Unknown error in alcDeviceResumeSOFT");
+      if (alcDeviceResumeSOFT != nullptr) {
+        BA_PRECONDITION_FATAL(impl_ != nullptr
+                              && impl_->alc_context != nullptr);
+        auto* device = alcGetContextsDevice(impl_->alc_context);
+        BA_PRECONDITION_FATAL(device != nullptr);
+        try {
+          g_core->platform->LowLevelDebugLog(
+              "Calling alcDeviceResumeSOFT at "
+              + std::to_string(g_core->AppTimeSeconds()));
+          alcDeviceResumeSOFT(device);
+        } catch (const std::exception& e) {
+          g_core->logging->Log(
+              LogName::kBaAudio, LogLevel::kError,
+              "Error in alcDeviceResumeSOFT at time "
+                  + std::to_string(g_core->AppTimeSeconds()) + ": "
+                  + g_core->platform->DemangleCXXSymbol(typeid(e).name()) + " "
+                  + e.what());
+        } catch (...) {
+          g_core->logging->Log(LogName::kBaAudio, LogLevel::kError,
+                               "Unknown error in alcDeviceResumeSOFT");
+        }
       }
 #endif
       last_started_playing_time_ = g_core->AppTimeSeconds();
@@ -922,10 +966,34 @@ void AudioServer::UpdateMusicPlayState_() {
   }
 }
 
+void AudioServer::ProcessDefaultDeviceChange_() {
+#if BA_OPENAL_IS_SOFT
+  if (should_reopen_) {
+    should_reopen_ = false;
+    auto* device = alcGetContextsDevice(impl_->alc_context);
+    assert(device);
+    if (alcReopenDeviceSOFT(device, nullptr, nullptr)) {
+      g_core->logging->Log(
+          LogName::kBaAudio, LogLevel::kInfo,
+          "alcReopenDeviceSOFT() succeeded after default output change.");
+    } else {
+      g_core->logging->Log(
+          LogName::kBaAudio, LogLevel::kError,
+          "alcReopenDeviceSOFT() failed after default output change.");
+    }
+  }
+#endif  // BA_OPENAL_IS_SOFT
+}
+
 void AudioServer::ProcessDeviceDisconnects_(seconds_t real_time_seconds) {
 #if BA_OPENAL_IS_SOFT
-  // If our device has been disconnected, try to reconnect it
-  // periodically.
+  // If our context device has been disconnected, try to reconnect it
+  // periodically. The Android back-end in particular uses uses this - with
+  // Android there is a single system device and plugging or unplugging
+  // headphones manifests as the device disconnecting and needing to be
+  // reset. For more info, see:
+  // https://github.com/kcat/openal-soft/issues/941
+
   auto* device = alcGetContextsDevice(impl_->alc_context);
   BA_PRECONDITION_FATAL(device != nullptr);
   ALCint connected{-1};
@@ -933,11 +1001,7 @@ void AudioServer::ProcessDeviceDisconnects_(seconds_t real_time_seconds) {
   CHECK_AL_ERROR;
   if (connected != 0) {
     last_connected_time_ = real_time_seconds;
-    // reconnect_fail_count_ = 0;
   }
-  // else {
-  // reconnect_fail_count_ = 0;
-  // }
 
   // Retry less often once we've been failing for a while.
   seconds_t retry_interval =
@@ -957,15 +1021,6 @@ void AudioServer::ProcessDeviceDisconnects_(seconds_t real_time_seconds) {
     auto result = alcResetDeviceSOFT(device, nullptr);
     CHECK_AL_ERROR;
 
-    // Log(LogLevel::kInfo, std::string("alcResetDeviceSOFT returned ")
-    //                          + (result == ALC_TRUE ? "ALC_TRUE" :
-    //                          "ALC_FALSE"));
-
-    // Check to see if this brought the device back.
-    // ALCint connected{-1};
-    // alcGetIntegerv(device, ALC_CONNECTED, sizeof(connected), &connected);
-    // CHECK_AL_ERROR;
-
     // If we were successful, clear out the wait for the next reset.
     // Otherwise plugging in headphones and then unplugging them immediately
     // will result in 10 seconds of silence.
@@ -981,32 +1036,6 @@ void AudioServer::ProcessDeviceDisconnects_(seconds_t real_time_seconds) {
         openalsoft_android_log_ += "DEVICE RESET CALL FAILED\n";
       }
     }
-
-    // If we're ever *not* immediately successful, flip on reporting to try
-    // and figure out what's going on. After that point we'll report subsequent
-    // if (connected == 0) {
-    //   report_reset_results_ = true;
-    // }
-    // if (report_reset_results_ && reset_result_reports_remaining_ > 0) {
-    //   reset_result_reports_remaining_ -= 1;
-    //   if (connected != 0) {
-    //     Log(LogLevel::kInfo,
-    //         "alcResetDeviceSOFT successfully reconnected device.");
-    //   } else {
-    //     Log(LogLevel::kError, "alcResetDeviceSOFT failed to reconnect
-    //     device.");
-    //   }
-    //   if (g_buildconfig.platform_android()) {
-    //     std::scoped_lock lock(openalsoft_android_log_mutex_);
-    //     Log(LogLevel::kWarning,
-    //         "------------------------"
-    //         " OPENALSOFT-RECONNECT-LOG-BEGIN ----------------------\n"
-    //             + openalsoft_android_log_
-    //         + "\n-------------------------"
-    //           " OPENALSOFT-RECONNECT-LOG-END -----------------------");
-    //     openalsoft_android_log_.clear();
-    //   }
-    // }
   }
 
   // If we've failed at reconnecting for a while, ship logs once.
@@ -1035,6 +1064,14 @@ void AudioServer::OnDeviceDisconnected() {
   Process_();
 }
 
+void AudioServer::OnDefaultDeviceChanged() {
+  assert(g_base->InAudioThread());
+  // All we do here is run an explicit Process_. This only saves us a half
+  // second or so over letting the timer do it, but hey we'll take it.
+  should_reopen_ = true;
+  Process_();
+}
+
 void AudioServer::Process_() {
   assert(g_base->InAudioThread());
   seconds_t real_time_seconds = g_core->AppTimeSeconds();
@@ -1042,9 +1079,17 @@ void AudioServer::Process_() {
 
   // Only do real work if we're in normal running mode.
   if (!suspended_ && !shutting_down_) {
-    ProcessDeviceDisconnects_(real_time_seconds);
+    // Handle audio output switching on Android (involves single system
+    // device disconnecting/reconnecting).
+    if (g_buildconfig.platform_android()) {
+      ProcessDeviceDisconnects_(real_time_seconds);
+    } else {
+      // Handle default audio device changing on multi-device systems.
+      // (plugging in headphones on a Mac, etc.)
+      ProcessDefaultDeviceChange_();
+    }
 
-    // Do some loading...
+    // Do some loading.
     have_pending_loads_ = g_base->assets->RunPendingAudioLoads();
 
     // Keep that available-sources list filled.
@@ -1153,12 +1198,6 @@ void AudioServer::FadeSoundOut(uint32_t play_id, uint32_t time) {
   sound_fade_nodes_.insert(
       std::make_pair(play_id, SoundFadeNode_(play_id, time, true)));
 }
-
-// void AudioServer::DeleteAssetComponent_(Asset* c) {
-//   assert(g_base->InAudioThread());
-//   c->Unload();
-//   delete c;
-// }
 
 AudioServer::ThreadSource_::ThreadSource_(AudioServer* audio_server_in,
                                           int id_in, bool* valid_out)
@@ -1394,8 +1433,8 @@ auto AudioServer::ThreadSource_::Play(const Object::Ref<SoundAsset>* sound)
 
   if (!g_base->audio_server->suspended_
       && !g_base->audio_server->shutting_down_) {
-    // Ok, here's where we might start needing to access our media... can't
-    // hold off any longer...
+    // Ok, here's where we might start needing to access our media; can't
+    // hold off any longer.
     (**source_sound_).Load();
 
     is_streamed_ = (**source_sound_).is_streamed();
@@ -1454,9 +1493,9 @@ void AudioServer::ThreadSource_::ExecPlay() {
     }
     audio_server_->streaming_sources_.push_back(this);
 
-    // Make sure stereo sounds aren't positional.
-    // This is default behavior on Mac/Win, but we enforce it for linux.
-    // (though currently linux stereo sounds play in mono... eww))
+    // Make sure stereo sounds aren't positional. This is default behavior
+    // on Mac/Win, but we enforce it for linux. (though currently linux
+    // stereo sounds play in mono... eww))
 
     bool do_normal = true;
     // In vr mode, play non-positional sounds positionally in space roughly
@@ -1477,10 +1516,12 @@ void AudioServer::ThreadSource_::ExecPlay() {
       throw Exception();
     }
 
-  } else {  // Not streamed
-    // Make sure stereo sounds aren't positional.
-    // This is default behavior on Mac/Win, but we enforce it for linux.
-    // (though currently linux stereo sounds play in mono... eww))
+  } else {
+    // Not streamed.
+    //
+    // Make sure stereo sounds aren't positional. This is default behavior
+    // on Mac/Win, but we enforce it for linux. (though currently linux
+    // stereo sounds play in mono... eww))
     if ((**source_sound_).format() == AL_FORMAT_STEREO16) {
       SetPositional(false);
       SetPosition(0, 0, 0);