+++ src/ballistica/shared/buildconfig/buildconfig_cmake.h
@@ -31,7 +31,7 @@
 #error Unknown processor architecture.
 #endif
 
-// #define BA_HAVE_FRAMEWORK_OPENAL 1
+// #define BA_USE_FRAMEWORK_OPENAL 1
 
 #elif __linux__
 
@@ -61,6 +61,13 @@
 
 #if !BA_HEADLESS_BUILD
 #define BA_ENABLE_AUDIO 1
+
+// NOTE: This needs to be flipped on to get fancy OpenAL stuff like auto
+// device switching, but it currently requires a newer OpenALSoft than our
+// build machines have. Should make this configurable.
+//
+// #define BA_OPENAL_IS_SOFT 1
+
 #define BA_ENABLE_OPENGL 1
 #define BA_SDL_BUILD 1
 #define BA_ENABLE_SDL_JOYSTICKS 1